requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
torch
torchvision
torchaudio
python-multipart
cryptography 
apscheduler
jinja2